fixing webpack loading
