+++ package.json
@@ -101,7 +101,6 @@
     "tslint-react-a11y": "1.1.0",
     "typedoc": "0.15.2",
     "typescript": "3.7.2",
-    "uglifyjs-webpack-plugin": "^2.2.0",
     "url-loader": "2.2.0",
     "webpack": "^4.29.6",
     "webpack-cli": "^4.9.1",
